Fixed the mistake caused to report "fatal:" from git.submodule.add: The reason was an absence on word "submodule" between "git" and "add" while trying to run "git submodule add <repo> <location>" ;)
  Currently "git submodule add ..." reports the following:
================================================================
A git directory for 'externals/sub' is found locally with remote(s):
  origin	D:\\TEMP\\GitRepo_sub
If you want to reuse this local git directory instead of cloning again from
  D:\TEMP\GitRepo_sub
use the '--force' option. If the local git directory is not the correct repo
or you are unsure what this means choose another name with the '--name' option.
================================================================
And this thing requires to be investigated.